Rename cloud-in to cloud-vrf-in
Building on taavi's earlier change it does make sense to rename
these.  'cloud-hosts-in' is the new name for the filter on traffic
from the cloud 'production' realm.  The filter on traffic from
the 'cloud' VRF on the switches (including cloud VM traffic and
traffic from other cloud-realm public IPs such as DNS) is just
called 'cloud-in' still, which is not terribly descriptive.

In terms of the CRs where it is applied I think 'cloud-vrf-in'
makes sense and is easy for us to know which is which at a glance.

Change-Id: I56fd53f888de9548c9eb32c21b805f6495601a93