WMCS: Update ACL / filter name for interface towards Cloud VRF
Omitted from previous patch I56fd53f888de9548c9eb32c21b805f6495601a93
which unfortunately means things are broken.

This patch references the new name of the filter on the devices it is
applied to so the config will generate properly.  Also updating
the descriptions in the YAML to be more sensible.

Change-Id: Iaf5af00d65b27ac66ee2ff3b061ccf9843cf8ebd