Use VC status to derive l3_switch variable and remove from YAML
We can set the l3_switch var at the top of our asw.conf now as we
have removed all the macros, and it can be fairly easily determined
based on the existing virtual_chassis data exposed from Homer (with
one exception for fundraising switches).

Tested and is a no-diff on all devices.

Change-Id: I39887353749f6f4cb3334fcc2d180b50b77108bf