User management: create new RO login class and allow to view logs
This patch defines a new login class 'wmf-ro' and grants members
rights to the same commands as the 'rancid' group it replaces.  This
is needed as the in-built "read-only" group does not allow certain
show commands to be run.

The command in system.conf is also needed to set the file permissions
on 'messages' files in /var/log/ so that all users can read them.

Tested with a temp user on one of our switches and works as expected.

Bug: T401378
Change-Id: Id8b29745ae1940b7a097b5cc506eb1d55ebc1ece