Srl_system: small fixes to make config apply and no diff
Adds the missing path for the host-name rpc.

Moves the _methods list inside the SrlOspf class otherwise the
OSPF config was not being generated.

Also and modifies the syslog function to take the hostnames and
port info from common.yaml and always use the port number (needed
for our default syslog on port 10514).

I did not factor this new data in common.yaml back into the Jinja
templates for JunOS yet.  There is some custom config there for
L2 ASWs which complicates it, will be easier to do when they are
gone.

Bug: T402577
Change-Id: I8004729efda63154960fc984be4060ba7aadfafa