Nokia: Save some vars to data{} dict so it only needs to be done once
Some small changes here.  Main one is to save some commonly-needed
items we work out back to the data{} dict so they can potentially
be used elsewhere, or we only need to create them once.

For instance the router-id var from the system0 IP address, getting
that now in nokia_asw.py, as well as the evpn bool var.

Likewise in network_instances.py saving the required list of vrfs
and vlans back to the data dict.  Also moved the function that
finds these to the end to prep for eventually moving some of it
to wmf-plugin.

These changes help moving the BGP stuff so we can build that
configuration for each network-inistance outside of the code that
creates the rest of the network-instance config.

Bug: T402577
Change-Id: I2e69e0b1f1eb75c3f9728228a836be14f44d7c0c