Redefine 'asns_mapping' to include additional bgp group metadata
Moving some elements out of hard-coded JunOS templates and into the
YAML structs in common.yaml so they can be also configured for the
Nokia switches.

Firstly replaces 'asns_mapping' with 'bgp_group_data', and then
moving the ASN info under a sub-key 'asns'.  Additionally moves
the ASNs for groups that hard been hard-coded into the YAML with
key name 'global' if there are not site-specific.

Also allows for new key - 'BFD' - to control whether BFD is enabled
for a particular group.

Bug: T402577
Change-Id: I1e1205c7fe0698fb565b5cbabe365d45e3a922a1