Nokia: EBGP configuration base build
Tested this off Netbox-next going through the various groups we need
for server BGP and all is good.  Some more work has to be done to
add the import/export policies but other than that it works.

Also added some statically defined EBGP peers, you can see the diff
from ssw1-d1-eqiad here:

https://phabricator.wikimedia.org/P83191

I expect there are improvements to be made here but hopefully this is
the broad approach we can take for this.

Bug: T402577
Change-Id: Icc267272101d898ea618c7096aa226815383e990