Nokia: extract system loopbacks IPs in nokia_asw.py for later
We need to use the system loopback IPs for various things in the
config (like sourcing sflow, dhcp relays etc.).

Moving the logic we had in the sflow part of system.py out and
instead extracting them and adding to the data dict in nokia_asw.py
so we can use them there and anywhere else.

Bug: T402577
Change-Id: Id1c57780506730a07bc4c07fbe98bcc92f72f164