Nokia: Add DHCP relay and IPv6 RA generation on IRB interfaces
IRB interfaces are the vlan interfaces on each top-of-rack switch
which act as the gateway for hosts on connected to them.  These
require two elements not found on other interfaces:

- DHCP relay in the IPv4 config to send traffic to our install server
- IPv6 router-advertisements configured in the v6 block

This patch adds those when the interface name matches.  Also removes
the 'evpn' bool var and instead passes the full config dict as we
need several elements from it for the above.

Bug: T402577
Change-Id: I904d24226de86cf009911e7b789124924241342c