Nokia: support mixing of L2 and L3 subinterfaces on SR Linux
Unlike Cisco/Juniper/Arista and most other vendors, interfaces in
SR Linux are not configured by default as either 'L2' or 'L3'.

Instead *everything* is a sub-interface of a parent physical, and
these can be tagged/untagged, and either L2 (attached to a vlan)
or L3 (with IP addressing on them) in whatever combination we
want.

This means we can use a regular dot1q sub-interface on the Nokia
side for BGP peering to our CRs, while still bridging the legacy
vlans for hosts doing BGP to the CRs.  Unlike on Juniper when
we need to do the BGP peering from an IRB interface and create a
vlan for it (yuck).

TL;DR this patch changes the code so that L2 sub-interfaces created
if a parent is set to Trunk in Netbox are not over-written if a L3
sub-interface is defined with the same parent port.  For instance
this netbox config (see ethernet-1/32):

https://netbox.wikimedia.org/dcim/devices/6352/interfaces/

Bug: T402577
Change-Id: Ia012c60eb2216b5eb1a183fe7a2a36b300a9c1d4