ssw1-d8-eqiad: add bgp peerings to CR and Juniper spines
Adds the EBGP peerings to cr2-eqiad from ssw1-d8-eqiad, as well as
the sessions from it to the Spine switches in E1 and F1.

Also adds the second LAG ESI definition which will connect from
ssw1-d1-eqiad and ssw1-d8-eqiad to asw2-d-eqiad for the migation.

Bug: T396063
Change-Id: Idf93234d34b0bba8220b5f4a7ba090082b5dfc03