Nokia: adjust how we load static YAML configs
This is probably a bad way to fix this, but it's simple for me.

Previously the routing-policy module used the load_yaml_config()
function from Homer's base config.py:

https://w.wiki/FXVF

This is unsuitable for loading direct configuration we want to push
to the Nokia devices, as it converts IP addresses and networks in
the source YAML to ipaddress.ip_address and ip_network objects.

That's great for loading data we need to use in the code, but for
direct config we need to keep things as strings, so the data can
be serialised as JSON and the box can understand it.

So here I have taken the step to change the get_static_config() in
the nokia __init__.py file, changing it from just getting the path
to also loading the YAML.  As the Homer code previously modifies
yaml.SafeLoader to parse IPs globally, it instead uses yaml.FullLoader
so the strings are retained.  This is not as bad as yaml.UnsafeLoader
but I think not ideal.  My own thinking is we are only parsing our
own saved YAML it should be ok.

Bug: T402577
Change-Id: I83877181f290269b0abeb3ef1624e1a7f02b4f62