eqiad: remove dedicate BGP trasnit to HE and replace with IX service
HE wanted to shut our PNI transit with them in Ashburn due to the
relatively small amount of traffic we send to them over it.

Instead they have offered to provdie us the same service over the
Equinix IXP.

This patch implements the change, which is already pushed to the
routers and working as it should.

Bug: T407008
Change-Id: I82a23f0980df632ef19fbdaceed5a6bb63e5989f