Nokia BGP: add function to get policy names based on BGP group name
Adds function to return the policy name to use for a BGP group (import
and export) based on the name of the group.

With Juniper traditionally we've been sort of ad-hoc with these, but
it makes more sense to standardise.  We can retroactively adjust our
Juniper configs to do the same in time.  Basically we will have
this policy:

- We look for a policy matching the group name, with '4' or '6' at
the end, representing an address-family specific policy
- If there is no address-fam specific policy we look for a policy
matching the group name (with no number at the end), and apply for
both address fams.
- If none found we set both policies to 'NONE' for safety

Also ammends the policy for pering to the CRs/other external networks.

Bug: T402577
Change-Id: I40e9ee2879714b4e4c4e6068a1b02783616708ef