Nokia DHCP-Relay: use IRB interface IP
Making dhcpd respond to a relayed request from a device when the
gi-addr is set to the system does not seem possible, at least
based on the testing we've done so far.

Ultimately we probably need the RFC3527 support feature Nokia I
believe are working on (same as we use for the Junipers).

Using the system loopback to relay the DHCP request is only needed
for our public Vlans, however, as they are the only ones where each
switch does not have a unique IP address on the host-facing vlan
subnet.  The per-rack vlans obviously have a unique IP as the gateway
does not exist on any other switch.  The row-wide we also allocate
a unique IP to each device as we can in the private space.  Only the
public has a single, shared anycast IP on every switch in the row.

Also adding the option 82 insertion commands, though it's not much
use to us in the current form (adds the L3/irb interface not L2 port)

TL;DR we will use the irb interface IP where we can, which will mean
the request will hit the DHCP server from an IP address in the range
it needs to assign from, which will work.  For the public vlan we
can probably leave our core routers connected to the subnet and
have them do the relay until Nokia have RFC3527 support.

Bug: T402577
Change-Id: I150117e8d757f157558d605af56849c345d6c982