Nokia: always set system cpm packet filter on devices
The config this new function produces is the same as default, which
we don't modify, so we'd not touched it in the code.

However as part of a test on one device I removed the filter, to see
if it was playing a role.  I forgot to re-add it, and the automation
did not either, which is a risk.

So we should always push this config to ensure the filter is in place.

Bug: T402577
Change-Id: I86b88ed97893e3d7da01872add6e7c9954b3ca6c