Eqiad C/D migration: move analytics1-c-eqiad GW to CR et-1/0/5
This moves the gateway IPs, router advertisement, OSPF conf etc for
the analytics1-c-eqiad vlan from the ae3 interface (which connects
to asw2-c-eqiad directly) to et-1/0/5 (which lands on ssw1-dX-eqiad
off each CR).

Effectively this moves the traffic path for this vlan from handing
off directly from the CR to Juniper virutal-chassis switch
(asw2-c-eqiad) to handing off to the new Nokia spines, which in
turn are connected to the old Juniper VC.  This brings the Nokia
into the path and allows us to begin migrating servers on the
vlan, before ultimately removing the VC switches completely.

Bug: T405579
Change-Id: I9940755a3eafaa3f4cbc79dcfd3bd1cb1f4dc4df