Eqiad C/D: Remove ESI-LAG config for Nokia connections to Juniper VCs
Once lvs1018 is reconfigured to only use it's primary uplink (in
row A) and not any cross-rack links to the Juniper VC stacks in
rows C and D we will have no more connections on the old Juniper
virtual-chassis switches there.

So we can decom the LAGs to the Nokias and shut the ports.  This is
mostly done in Netbox which I can take care of, but this patch
should be merged after to remove the ESI ID for the EVPN bit.

Bug: T411781
Change-Id: Icc7df57753564e5ba8a12fbc25b70520b95e94af