Nokia sflow: disable on spine switch access/trunk ports
After fixing the sflow issue with flow data not being allowed from
the spine/leaf loopback ranges we can now see the Nokia spines in
eqiad rows C/D are the top exporters of flows, from their configured
uplinks to the core routers.

This traffic should not be included in sflow, as we only want to
have sampling enabled on server-facing ports to catch internal
traffic, and only catch it once.

Tiny bit messy to do this but this seemed the best way.

Change-Id: I888e1fb9ed073cef9dc080d64dbfc15414b29ce1