Nokia: add support for SR-Linux v25 or v24
Adds a var in to specify the SR Linux version to build a
device configuration for.  The var can be set at global/site/
device level, currently set at site level for eqiad and codfw.

Not exactly perfect but may get us over the short-term issue as
we consider the better way to deal with the issue longer term.

Bug: T412157
Change-Id: I606552566fbf88b7b73c1e7c17145a28fe467b04