Nokia ESI-LAG: Adjust module to fully remove when last LAG deleted
With the config as it was if no ESI-LAGs were defined for a device
it would not push any config under the 'system' path for them.

That caused it to not remove that part of the config when all
active LAGs were deleted from the eqiad Spine switches in C/D,
leaving the configuration under 'system/protocols' in place.

Instead always return config, an empty dict in the case there
are none defined.

Change-Id: I09377542af0d655159bc9841b06a966c8d447ea5