Nokia: set FEC mode explicity for all 100G links on v25
It seems some default changed in SR-Linux v25, so a 100G CWDM4 link
connected to another Nokia requires the FEC mode to be set
explicitly or it won't come up.

Tested this and it's a no-op everywhere (adds it on lswtest-d8 as
needed).  When we have everything on v25 we can review what we need,
perhaps the defaults will be ok without setting it once all is
upgraded.

Bug: T412157
Change-Id: Ic691e367096aa9c9d3d2a775119ecf00477102ba